Handle twitter auth in same window (#46)
* handle sign-in in new tab

* extract copy to component

* remove `react-new-window`

* handle sign in on same tab
