+++ package.json
@@ -32,7 +32,6 @@
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-jazzicon": "^1.0.3",
-    "react-new-window": "^0.2.1",
     "ua-parser-js": "^1.0.2"
   },
   "devDependencies": {
